requirements.txt
Django>=4.2.10
djangorestframework>=3.14.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
pycryptodome>=3.20.0
django-cors-headers>=4.3.1
python-dateutil==2.9.0
